% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{assign_genes_to_paths}
\alias{assign_genes_to_paths}
\title{Assign Genes to Paths}
\usage{
assign_genes_to_paths(m, minimal_edge_set, ...)
}
\arguments{
\item{m}{GeneMatrix object representing GO terms}

\item{minimal_edge_set}{a network which is a weighted directed minimal spanning tree of gene ontologies}

\item{...}{Extra arguments passed on to threshold_sets}
}
\value{
a data_frame which contains which terminal gene set each gene is assigned to
}
\description{
Assign genes to a GO hierarchy which has the greatest signal
}

