\name{CompareTopSets}
\alias{CompareTopSets}
\title{Create a graph comparing the top n columns (in terms of association with y)}
\usage{
CompareTopSets(m, n = 9, ...)
}
\arguments{
  \item{m}{GeneMatrix object}

  \item{n}{Number of genes to compare}

  \item{...}{Additional arguments to be passed to CompareY}
}
\description{
Create a graph comparing the top n columns (in terms of
association with y)
}

