% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CompareY.R
\name{ExtractGenes}
\alias{ExtractGenes}
\title{Return all genes, along with the saved metric in each of the given sets}
\usage{
ExtractGenes(m, columns, char_limit = NULL, add_count = FALSE,
  overall = FALSE)
}
\arguments{
\item{m}{a GeneMatrix object}

\item{columns}{vector of columns (e.g. gene sets) whose genes should be extracted}

\item{char_limit}{optionally shorten the Term field to the desired length}

\item{add_count}{whether the count within each set should be added to the
Term field (like "Metabolism (360)")}
}
\description{
Return all genes, along with the saved metric in each of the given sets
}

