% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{find_minimal_rooted_tree}
\alias{find_minimal_rooted_tree}
\title{Find Minimal Rooted Tree}
\usage{
find_minimal_rooted_tree(nodes, edges)
}
\arguments{
\item{nodes}{data_frame significance of gene sets}

\item{edges}{data_frame go_id1 are parents of go_id2}

\item{ancestors}{data_frame go_id1 are ancestors of go_id2}
}
\value{
an igraph object of the gene set minimal spanning tree
}
\description{
Removes all loops so that gene sets can be nested within one another: creates an optimal directed acyclic graph
pointing from general GO terms down to specific ones.
}

