% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treemap.R
\name{GenerateTreemap}
\alias{GenerateTreemap}
\title{Generate Treemap}
\usage{
GenerateTreemap(m, edges = NULL, ...)
}
\arguments{
\item{m}{GeneMatrix object representing GO terms}

\item{...}{Extra arguments passed on to threshold_sets}
}
\value{
Generates a treemap hierarchy
Treemap GO Plot

Generate a treemap plot using the minimal spanning GO tree with genes assigned to individual GO terms.

Default color based on GO significance and effect sizes of individual genes
}
\description{
Visualize genes nested within specific and general GO sets using treemaps
}
\details{
Each gene should maximize the absolute sum of relevant node weights but every GO category will
not be included since a gene can only be assigned to a single ancestry. A small GO category may
fall into multiple ancestral paths some which are parents of other meaningful categories and others
that are dead-ends.  Should favor attachment to pathways with the most informative children.
To allow this to occur, three edge weights are used
1) the path with the highest overall absolute score * number of genes affected is chosen
2) ancestors of LASSO nodes all get a smaller score: [min(abs(beta))]*0.01*N_lasso_children
- applied to edge where ancestors are children
3) all remaining edges that are not predictive recieve a tiny weight so that they are not filtered: [min(abs(beta))]*1e-4
}

