% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GOMembershipMatrix.R
\name{get_ancestry_matrix}
\alias{get_ancestry_matrix}
\title{Build an offspring matrix of GO terms}
\usage{
get_ancestry_matrix(sets)
}
\arguments{
\item{terms}{IDs of GO terms that should be included in the ancestry matrix}
}
\description{
A sparse binary Matrix object with one row and column for each pair
of GO terms provided, where each row represents an ancestor and each column
represents a descendant, with 1 marking ancestor/descendant pairs.
}

