% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{GetEdgesTable}
\alias{GetEdgesTable}
\title{Create a data frame of edges from a GO term matrix}
\usage{
GetEdgesTable(m, sets = NULL, ancestors = TRUE, ontology = c("BP", "MF",
  "CC"), ...)
}
\arguments{
\item{m}{GeneMatrix object representing GO terms}

\item{sets}{Gene sets to use in the network. If NULL, use
\code{\link{ThresholdSets}} to find them}

\item{ancestors}{Whether to include all ancestors of significant
in the matrix}

\item{ontology}{Which of BP, MF, and CC to use (default all three)}

\item{...}{Extra arguments passed on to threshold_sets}
}
\value{
An igraph
}
\description{
Create a data frame of edges from a GO term matrix
}

