% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOMembershipMatrix.R
\name{GOMembershipMatrix}
\alias{GOMembershipMatrix}
\title{Construct a new gene set membership matrix}
\usage{
GOMembershipMatrix(annotations, ontology = c("BP", "MF", "CC"),
  evidence = NULL, min_size = 1, max_size = Inf, ancestors = TRUE,
  chosen_genes = NULL)
}
\arguments{
\item{annotations}{A Go3AnnDbBimap object, typically from a Bioconductor Annotation
package. For example, org.Hs.egGO from org.Hs.eg.db for human genes.}

\item{ontology}{A vector of ontologies to include; should be a subset
of c("BP", "MF", "CC")}

\item{evidence}{A vector of GO evidence codes to include}

\item{min_size}{Minimum size of a gene set to be included}

\item{max_size}{Maximum size of a gene set to be included}

\item{ancestors}{Whether a gene included in a gene set should also be included
as being in all ancestors of that gene set (default TRUE)

The GO annotation maps typically come from the Bioconductor AnnotationDB packages.
A couple of notable examples are:

Homo sapiens: org.Hs.egGO
S. cerevisiae: org.Sc.sgdGO
E coli K12: org.EcK12.egGO

You can restrict the sets to the BP (Biological Process), MF (Molecular Function),
or CC (Cellular Compartment) ontologies (by default all are included).}
}
\description{
Construct a gene set membership matrix (one row per gene, one column per
gene set, with 1 indicating membership) from a GO annotation map.
}
\examples{

# yeast membership matrix
library(org.Sc.sgd.db)
mm <- GOMembershipMatrix(org.Sc.sgdGO, min_size = 5, max_size = 250)

# restrict to Biological Process ontology:
mm <- GOMembershipMatrix(org.Sc.sgdGO, ontology = "BP", min_size = 5, max_size = 250)

# human membership matrix
library(org.Hs.eg.db)
mm <- GOMembershipMatrix(org.Hs.sgdGO, min.size = 5, max.size = 250)

}

