% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAssociation.R
\name{mean_difference}
\alias{mean_difference}
\title{Compute mean differences between "in a set" and "not in a set"}
\usage{
mean_difference(m)
}
\arguments{
\item{m}{GeneMatrix object}
}
\description{
For each set in a GeneMatrix, compute the difference in y between
"in the set"
and "outside the set". It will be set in the MeanDifference column
}

