% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{PlotNetwork}
\alias{PlotNetwork}
\title{Plot a network from an igraph representation}
\usage{
PlotNetwork(g, algorithm = "kk", arrow = grid::arrow(length =
  grid::unit(0.1, "inches")))
}
\arguments{
\item{g}{An igraph object, generally computed by \link{\code{GenerateNetwork}}}

\item{algorithm}{Algorithm to use for layout}

\item{arrow}{Arrow}
}
\description{
Return a ggplot2 object representing a network of sets.
This is meant to be a sensible default for plotting
GO networks, it will not cover all cases. We recommend learning
to use ggraph or another igraph plotting functionality.
}

