% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestAssociation.R
\name{ThresholdSets}
\alias{ThresholdSets}
\title{Return a vector of genes that pass a threshold for includions}
\usage{
ThresholdSets(m, alpha = 0.05, method = "fdr", ...)
}
\arguments{
\item{m}{GeneMatrix object}

\item{alpha}{Threshold for (corrected) p-values, not used in LASSO}

\item{method}{Method, passed to \code{\link{p.adjust}}, or "qvalue"
to use qvalue. Not used in LASSO}

\item{...}{Extra arguments to pass on to correction method}
}
\description{
Return a vector of genes that pass a threshold of "significance"
or other kind of inclusion. For tests with a p-value (t-test,
Wilcoxon, hypergeometric), this is FDR-controlled p-values.
For LASSO, this is all cases where beta1sd != 0.
}
\seealso{
\link{p.adjust}, \link{p.adjust.methods}
}

