% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{trim_ellipses}
\alias{trim_ellipses}
\title{Trim the length of a character vector, adding ellipses to trimmed elements}
\usage{
trim_ellipses(x, width)
}
\arguments{
\item{x}{character vector}

\item{width}{length of string to trim to}
}
\description{
This utility function is used to shorten column names so they can appear
on a figure.
}

