% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOMembershipMatrix.R
\name{get_ancestry_matrix}
\alias{get_ancestry_matrix}
\title{Build an offspring matrix of GO terms}
\usage{
get_ancestry_matrix(terms, ontology = c("BP", "MF", "CC"),
  type = "OFFSPRING", upward = TRUE, tbl = FALSE)
}
\arguments{
\item{terms}{IDs of GO terms that should be included in the ancestry matrix}

\item{ontology}{Ontologies to use}

\item{type}{Either OFFSPRING (default), CHILDREN, ANCESTOR, or PARENT}
}
\description{
A sparse binary Matrix object with one row and column for each pair
of GO terms provided, where each row represents an ancestor and each column
represents a descendant, with 1 marking ancestor/descendant pairs.
}

